one process per baker (#517)
* one process per baker

When baking for several addresses, today we instantiate only one baker
process. This is unsupported and bad, because it does things
sequentially. The supported way to bake for several addresses is to run
one baking process per address. This PR changes tezos-k8s to do that.

This is a second, less intrusive attempt after feedback from Aryeh.

* helm computes the maximum number of baking processes that any pod in
  the statefulset would need
* instantiates the max number of bakers for each pod (no choice here..)
* the baker entrypoint figures which baker it must bake for thanks to a
  new BAKER_INDEX env var. If it must idle, it enters an infinite
  sleeping loop.

This works when using `bake_using_accounts`. I verified that I did not
break `bake_using_account`.

I verified by:
* creating a private chain with 3 bakers using mkchain
* assigning two bakers to the first node and one baker to the second

I also ran mkchain with no parameters to ensure it still works.

Various cleanups:

* remove a octez v13 specific catch (must pass toggle even when vote
  file absent) in the baker entrypoint
* remove the DAEMON env var in baker (was used to disambiguate between
  baker and endorser)

Random comment:

I don't think it's great that the logic to pick the baking accounts is
written in config-generator.sh instead of py, but I'm not changing this
here.

* reinstate DAEMON env var per @elric1 demand