Set fitness to 1 instead of -1 at activation
A bug was uncovered during mumbai activation, where nodes were unable to
activate and were crashing.

The [fix](https://gitlab.com/tezos/tezos/-/merge_requests/7462/diffs)
broke dailynet because it is no longer possible to activate with a
fitness of -1 (which is how we have been activating chains since the
beginning of tezos-k8s).

Replacing this value to 1 fixes the issue for dailynet and also does not
appear to be causing any issue with v15.1.