Persist identity and peers when using local storage
Local storage is great because it's faster, and we can download tarballs
in minutes at every boot and have tolerable downtimes.

But at each boot, the node regenerates its identity. We already have a
mechanism to pre-populate identities in the node yaml, but it's
cumbersome.

Instead, with this PR, we create a mini-pvc of 1Gi to store node
identities, whenever local storage is in use.

This way, local storage nodes always boot quickly, except the first time
they boot, and there is no configuration burden. It is also possible to
scale up the number of nodes without having to prepare identities.

I tested private chain and ghostnet node, in both local and non-local
storage. It behaves as intended. I also tried to pre-populate an
identity in the values file in local storage and confirmed that it is
not broken.