per-baker toggle vote customization
This allows a baking tezos-k8s setup to customize votes per baker.

Until now, the per-baker votes are set under the `protocols:` object of
values.yaml. The votes are listed per prototol and apply to all bakers.

You can now pass a similar `protocols:` object to each account under the
`accounts:` object of values.yaml. It will override the global config.

I tried on minikube with 2 bakers, with or without customization of the
votes. I observed in the baker logs that the toggles were as expected.