Add lock to tezos sidecar
When the node has a lot of contention due to a lot of RPC requests, I
noticed the following:

* node keeps processing new blocks
* node does not reply to any RPC request even /version, even locally
* doing netstat on the node shows a lot of connections from localhost
  (sidecar connections)
* sometimes the node recovers on its own, sometimes not

I am wondering if the sidecar is not making things worse in this
context, deadlocking the node even more. So I am proposing that the
sidecar will only make one http request to the node at a time using
multiprocessing.Lock

I also asked Nomadic engineers what they think about this.