fix tezos pod crash when config already exists (#565)
* fix tezos pod crash when config already exists

Today I found a baking node in this state: the init container was in a
failure loop. `config-init` had the following message:

```
Writing custom configuration for public node

octez-node: Error:

              Pre-existing configuration file at
/etc/tezos/data/config.json, use `reset`.
```

Addressing this by ensuring config-init always works, despite
pre-existing data. This does not solve the root cause (unknown) but
ensures that the tezos pod starts in this scenario.

This is not usual; config-volume is an emptydir so when the pod restart,
it should be empty. It could be that this specific cloud provider does
not clear the config dir upon pod reboot, or the init-containers re-run
in the pod when config-volume is not empty.

* fix syntax

* fix tests