chain activation: avoid misfiring
when activate job gets a successful RPC response for the first time, it
should not repeat the query; instead, just parse the output again and
check whether the level is zero.

This way, if the second rpc query fails, it will just restart the job
instead of passing silently, resulting in a non-activated chain stuck at
level 0, as happened on Mondaynet today.